Add CI matrix (linux/darwin/windows, x86_64 + aarch64) (#7)
* ci: add GitHub Actions matrix for linux/darwin/windows on x86_64 + arm64

Build and run the test suite on six GitHub-hosted runners:
linux x86_64 (ubuntu-latest) and aarch64 (ubuntu-24.04-arm), darwin
x86_64 (macos-13) and aarch64 (macos-14), windows x86_64
(windows-latest) and aarch64 (windows-11-arm). CI uses the vendored
third-party submodules (no Conan), so checkout is recursive.

Make CMakeLists cross-platform: the GNU/Clang-only flags (-Wall -Werror
-pedantic, -ffunction-sections, -fpic, the -f* optimization flags, and
the -Wl,--gc-sections/-dead_strip link options) are now guarded behind
NOT MSVC, with MSVC using /W4 /WX for the strict-warning bar.


* doc: add CI status badge to README


* ci: fix MSVC /WX warnings (size_t narrowing, CRT fopen deprecation)

The Windows runners surfaced warnings-as-errors not seen on GNU/Clang:
- C4267 size_t->int narrowing: the example length locals were int and
  the C++ encode_end() returned int. Use size_t in the examples and
  std::streamsize for encode_end (consistent with encode()/decode()).
- C4996 CRT 'fopen is unsafe': define _CRT_SECURE_NO_WARNINGS for MSVC;
  we use standard C I/O deliberately.

Verified GNU/Clang build stays -Werror -pedantic clean (30/30 tests).


* ci: fix _CRT_SECURE_NO_WARNINGS define (shim needs -D prefix)

add_compile_definitions is overridden by this file's compat shim to call
add_definitions(${ARGV}), so a bare _CRT_SECURE_NO_WARNINGS was passed to
cl as a source-file arg (C1083) for every target. Include the -D prefix.


* ci: per-platform workflows + reusable build, macos-15-large for Intel

Restructure to match the QVXLabs/xsd-tools convention: one workflow per
platform (ci-<platform>.yml) calling a reusable _build.yml, so each
platform gets its own status badge. Replaces the single ci.yml matrix.

Switch Intel macOS from macos-13 (the last free hosted Intel image, now
scarce and slow to schedule) to macos-15-large, which is billed but
actually available — same choice xsd-tools made.

README now shows a per-platform badge for each of the six targets.


* Tighten verbose CMakeLists comments to one or two lines


---------

+++ examples/c-example2.c
@@ -27,13 +27,13 @@ encoding and decoding the data.
 void encode(FILE* inputFile, FILE* outputFile)
 {
 	/* set up a destination buffer large enough to hold the encoded data */
-	int size = SIZE;
+	size_t size = SIZE;
 	char* input = (char*)malloc(size);
 	char* encoded = (char*)malloc(2*size); /* ~4/3 x input */
 	/* we need an encoder and decoder state */
 	base64_encodestate es;
 	/* store the number of bytes encoded by a single call */
-	int cnt = 0;
+	size_t cnt = 0;
 	
 	/*---------- START ENCODING ----------*/
 	/* initialise the encoder state */
@@ -62,13 +62,13 @@ void encode(FILE* inputFile, FILE* outputFile)
 void decode(FILE* inputFile, FILE* outputFile)
 {
 	/* set up a destination buffer large enough to hold the decoded data */
-	int size = SIZE;
+	size_t size = SIZE;
 	char* encoded = (char*)malloc(size);
 	char* decoded = (char*)malloc(1*size); /* ~3/4 x encoded */
 	/* we need an encoder and decoder state */
 	base64_decodestate ds;
 	/* store the number of bytes encoded by a single call */
-	int cnt = 0;
+	size_t cnt = 0;
 	
 	/*---------- START DECODING ----------*/
 	/* initialise the encoder state */
